requirements.txt
numpy
matplotlib
opencv-python
scikit-image
imageio
scipy
Pillow
torch
torchvision
transformers
diffusers
accelerate
